+++ ATmega2560_STS/ATmega2560_STS/Lys.h
@@ -12,11 +12,10 @@
 #include "Master.h"
 
 // Funktionsprototyper
-
 void adc_init(void);                         // Initialiser ADC
 void getSteps(); // Beregn motorbevægelser
 uint16_t calculate_sun_position(uint16_t sensor1, uint16_t sensor2, uint16_t correctionSize); // Beregn trin
-int get_light_intensity();
+int16_t get_light_intensity();
